refactor: removed user service dependencies from feed service
Splitting the monolithic application originally covering Feed and User services,
I have removed all dependencies from User Service
