+++ udacity-c3-restapi-feed/src/config/config.ts
@@ -9,15 +9,8 @@ export const config = {
     "aws_profile": process.env.AWS_PROFILE,
     "aws_media_bucket": process.env.AWS_BUCKET
   },
-  "prod": {
-    "username": "",
-    "password": "",
-    "database": "udagram_prod",
-    "host": "",
-    "dialect": "postgres"
-  },
   "jwt": {
     "secret": process.env.JWT_SECRET
   }
 
-}
+};
